Orders implemented

+++ eStoreRest/src/main/java/com/ss/estore/service/UserService.java
@@ -7,7 +7,7 @@ import com.ss.estore.model.User;
 
 public interface UserService {
 	public void add(User user);
-	public User fetch(int userId);
+	public User fetch(long userId);
 	public void addAll(Collection<User> stores);
 	public User deleteUser(long userId);
 	public List<User> listAll();
